mobile-engage-sdk/src/main/java/com/emarsys/mobileengage/iam/lifecycle/IamActivityLifecycleListener.java
package com.emarsys.mobileengage.iam.lifecycle;

import android.app.Activity;
import android.app.Application;
import android.os.Bundle;

public class IamActivityLifecycleListener implements Application.ActivityLifecycleCallbacks {

    private Activity currentActivity;

    public Activity getCurrentActivity() {
        return currentActivity;
    }

    @Override
    public void onActivityCreated(Activity activity, Bundle savedInstanceState) {

    }

    @Override
    public void onActivityStarted(Activity activity) {

    }

    @Override
    public void onActivityResumed(Activity activity) {
        currentActivity = activity;
    }

    @Override
    public void onActivityPaused(Activity activity) {
        if (currentActivity == activity) {
            currentActivity = null;
        }
    }

    @Override
    public void onActivityStopped(Activity activity) {

    }

    @Override
    public void onActivitySaveInstanceState(Activity activity, Bundle outState) {

    }

    @Override
    public void onActivityDestroyed(Activity activity) {

    }
}
